Add a location where one can prevent SYS access
Accessing the SYS schema is unwieldingly slow. Really. Nor does it have
a way to get the primary key info for a table. If one works with a stable
database design, having something available that returns you that info
*instantly*, one can gain a lot. Example below.

If the custom function does not return a key, the "normal" approach will
be used.

package My::U2000;

my %keys = (
    my_schema => {
        foo => [qw( code stamp )],
        },
    );

sub get_key
{
    my ($schema, $table) = @_;
    my $keys = $keys{$schema}{$table} || [];
    # research case-insensative or without schema if empty
    return @$keys
    } # get_key